Fixed #44 EASY_DEBUGGING constant is implemented.
