+++ MonoDevelop.XUnit/MonoDevelop.UnitTesting.XUnit/IRemoteEventListener.cs
@@ -1,0 +1,53 @@
+//
+// ExternalTestRunner.cs
+//
+// Author:
+//   Lluis Sanchez Gual
+//
+// Copyright (C) 2005 Novell, Inc (http://www.novell.com)
+//
+// Permission is hereby granted, free of charge, to any person obtaining
+// a copy of this software and associated documentation files (the
+// "Software"), to deal in the Software without restriction, including
+// without limitation the rights to use, copy, modify, merge, publish,
+// distribute, sublicense, and/or sell copies of the Software, and to
+// permit persons to whom the Software is furnished to do so, subject to
+// the following conditions:
+// 
+// The above copyright notice and this permission notice shall be
+// included in all copies or substantial portions of the Software.
+// 
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+// EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+// MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+// NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+// LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+// OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+// WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+//
+
+using System;
+using System.IO;
+using System.Collections.Generic;
+
+using MonoDevelop.Core;
+using MonoDevelop.Core.Execution;
+using System.Threading.Tasks;
+using MonoDevelop.XUnit;
+
+namespace MonoDevelop.UnitTesting.XUnit.External
+{
+
+	public interface IRemoteEventListener
+	{
+		void OnTestCaseStarting(string id);
+		void OnTestCaseFinished(string id);
+		void OnTestFailed(string id,
+			decimal executionTime, string output, string[] exceptionTypes, string[] messages, string[] stackTraces);
+		void OnTestPassed(string id,
+			decimal executionTime, string output);
+		void OnTestSkipped(string id,
+			string reason);
+	}
+	
+}
